feat(tas): zero-cost CONT backlog reset at the splice (replaces decel) — RE-located prev_time
Located the game clock via CheatEngine MCP: the tick scheduler (exe+0x25C81) runs
with ebp = the clock object (vtable 0x46D6EC, holds a date string); the time
accumulator ("prev_time") is the float at [ebp+0x0C], updated by
prev += ticks*tick_advance (exe+0x25E2B). Constants: [0x46DB08]=0.01 (tick_advance,
cave5-scaled for speed), [0x46DB0C]=100.0 (=1/0.01, NOT scaled) — so the raw
__ftol demand (esi) = elapsed*100 = (now-prev)/native.

Fix: at the splice, cave2 sets cont_reset_pending; on the next cave5 tick we
advance the accumulator to "now" via [ebp+0x0C] += raw_demand * NATIVE(0.01),
WITHOUT processing the backlog ticks (ctx.esi stays capped). prev jumps to wall —
no recorded burst, no huge-dt jump — so the resume is frame-exact AND the catch-up
keeps full speed (the ~350ms decel is gone). Native (not the speed-scaled
tick_advance) makes it speed-independent, so it should hold at 96x/128x too.

Removed the deceleration. Shared struct gains cont_reset_pending u32 (size pin
1_647_240→1_647_248, input_log offset unchanged). DLL rebuilt; deploy + run
fe10065-cont to validate.

+++ TAS/tas_ui/src/history_store_v2.rs
@@ -1105,7 +1105,7 @@ mod tests {
         let mut next_id = 1u64;
 
         for _ in 0..200 {
-            let n = 3 + rng.below(6) as u64;
+            let n = 3 + rng.below(6);
             let ids: Vec<u64> = (0..n).map(|_| { let id = next_id; next_id += 1; id }).collect();
             let model: Vec<StoredEntry> =
                 ids.iter().map(|&id| entry(id, &format!("e{}", id), false, 2)).collect();
